+++ src/components/Main/Company/Table/RowDetail/DetailView/DetailView.jsx
@@ -20,6 +20,7 @@ import {
 } from "./DetailView.utils.js";
 import { CompanyContext } from "../../../../../../store/company-context.jsx";
 import {
+  apiCallToCreateEntity,
   apiCallToUpdateEntity,
   apiCallToDeleteEntity,
 } from "../../../../../../api/Api.jsx";
@@ -27,7 +28,6 @@ import {
 const DetailView = memo(function DetailView({
   entity,
   entityName,
-  fieldName,
   updateRowHandler,
   deleteRowHandler,
   rowIndex,
@@ -35,7 +35,7 @@ const DetailView = memo(function DetailView({
   const [response, setResponse] = useState(null);
   const [showParents, setShowParents] = useState(false);
   const [showChilds, setShowChilds] = useState(false);
-  const {parents} = useContext(CompanyContext);
+  const { parents } = useContext(CompanyContext);
 
   const timer = useRef();
 
@@ -59,24 +59,40 @@ const DetailView = memo(function DetailView({
     });
 
     const formData = Object.fromEntries(fd.entries());
-    const updatedEntity = {...formData, ...parents}
-    
+    const updatedEntity = { ...formData, ...parents };
+
     disabledElements.forEach((element) => (element.disabled = true));
 
     let resData;
 
     try {
       if (action === "save") {
-        resData = await apiCallToUpdateEntity(
-          entityName.toLowerCase(),
-          updatedEntity
-        );
-        updateRowHandler(rowIndex, formData);
+        if (entity.id === 0) {
+          updatedEntity.id = null;
+          resData = await apiCallToCreateEntity(
+            entityName.toLowerCase(),
+            updatedEntity
+          )
+            .then((resData) => {
+              formData.id = resData.split(":")[1];
+              updateRowHandler(0, formData);
+              return resData.split(":")[0];
+            });
+        } else {
+          resData = await apiCallToUpdateEntity(
+            entityName.toLowerCase(),
+            updatedEntity
+          ).then(updateRowHandler(rowIndex, formData));
+        }
       } else {
-        resData = await apiCallToDeleteEntity(
-          entityName.toLowerCase(),
-          updatedEntity
-        );
+        if (entity.id === 0) {
+          resData="Canceled"
+        } else {
+          resData = await apiCallToDeleteEntity(
+            entityName.toLowerCase(),
+            updatedEntity
+          );
+        }
       }
     } catch (error) {
       resData = error;
@@ -88,10 +104,11 @@ const DetailView = memo(function DetailView({
 
     timer.current = setTimeout(() => {
       setResponse(null);
-      if (action === "delete" && !resData instanceof Error) {
-        deleteRowHandler(fd.get('id'));
+      if (action === "delete" && !(resData instanceof Error)) {
+        const id = entity.id === 0 ? 0 : fd.get("id");
+        deleteRowHandler(id);
       }
-    }, 5000);
+    }, 3000);
   }
 
   function showParentHandler() {