+++ src/components/Main/Company/Table/RowDetail/DetailView/DetailView.utils.js
@@ -17,7 +17,6 @@ export function determineUniqueFieldName(entityName) {
 
 function determinateFieldType(field) {
   const numberFields = [
-    "id",
     "apartmentNumber",
     "capacity",
     "carbonLimit",
@@ -61,7 +60,7 @@ export function entityMappingHandler(entity, entityName) {
             name={key}
             $align="center"
             disabled={determinateFieldDisabled(key, entityName)}
-            placeholder={entity[key]}
+            placeholder={entity[key] ? entity[key] : key}
             $color={determinateFieldDisabled(key, entityName) ? "" : entityName}
             type={determinateFieldType(key)}
           ></$RowInputField>