+++ src/components/Main/Company/Table/Table.jsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useState } from "react";
+import { useContext, useEffect, useRef, useState } from "react";
 import {
   flexRender,
   getCoreRowModel,
@@ -13,7 +13,7 @@ import {
   determineFieldNameHandler,
 } from "./Table.utils.js";
 import { CompanyContext } from "../../../../store/company-context.jsx";
-import { apiCallToGetListOfEntities } from "../../../../api/Api.jsx";
+import { apiCallToGetListOfEntities, apiCallToGetEntityTemplate } from "../../../../api/Api.jsx";
 import { $TableContainer, $TableBox, $Table } from "./Table.styles.jsx";
 import Header from "./Header/Header.jsx";
 import Search from "./Search/Search.jsx";
@@ -21,15 +21,14 @@ import Pagination from "./Pagination/Pagination.jsx";
 import RowDetail from "./RowDetail/RowDetail.jsx";
 
 export default function Table() {
-  const [tableData, setTableData] = useState({
-    data: [{}],
-    columns: [{ accessorKey: "na", id: "na", header: "na" }],
-  });
+  const [header, setHeader] = useState([{ accessorKey: "na", id: "na", header: "na" }]);
+  const [rows, setRows] = useState([{}]);
   const { selected } = useContext(CompanyContext);
+  const template = useRef({});
 
   const table = useReactTable({
-    data: tableData.data,
-    columns: tableData.columns,
+    data: rows,
+    columns: header,
     getCoreRowModel: getCoreRowModel(),
     getSortedRowModel: getSortedRowModel(),
     getFilteredRowModel: getFilteredRowModel(),
@@ -42,12 +41,24 @@ export default function Table() {
   });
 
   useEffect(() => {
-    for (const row of table.getRowModel().rows) {
-      if (row.getIsExpanded()) {
-        row.toggleExpanded();
-      };
+    shrinkTableRowsHandler();
+
+    async function fetchData() {
+      try {
+        await apiCallToGetEntityTemplate(selected.toLowerCase()).then((resData) => {
+          const columns = getTableColumns(resData, selected);
+          template.current = resData;
+          setHeader(_prevValue => columns);
+        });
+      } catch (error) {
+        // TO DO error
+      }
     }
 
+    fetchData();
+  }, [selected]);
+
+  useEffect(() => {
     async function fetchData() {
       try {
         await apiCallToGetListOfEntities(
@@ -56,9 +67,7 @@ export default function Table() {
           "",
           true
         ).then((resData) => {
-          const columns = getTableColumns(resData[0], selected);
-          const table = { data: resData, columns: columns };
-          setTableData((_prevData) => table);
+          setRows((_prevData) => resData);
         });
       } catch (error) {
         //TODO
@@ -66,14 +75,29 @@ export default function Table() {
     }
 
     fetchData();
-  }, [selected]);
+  }, [header]);
+
+  const shrinkTableRowsHandler = () => {
+    for (const row of table.getRowModel().rows) {
+      if (row.getIsExpanded()) {
+        row.toggleExpanded();
+      };
+    }
+  }
+
+  const createRowHandler = () => {
+    setRows((_prevData) => {
+      const data = [template.current, ..._prevData];
+      return data;
+    });
+  };
 
   const updateRowHandler = (rowIndex, newData) => {
-    setTableData((_prevData) => {
-      const data = _prevData.data.map((row, index) =>
+    setRows((_prevData) => {
+      const data = _prevData.map((row, index) =>
         index === rowIndex ? newData : row
       );
-      return { ..._prevData, data };
+      return data;
     });
   };
 
@@ -83,9 +107,10 @@ export default function Table() {
         row.toggleExpanded();
       };
     }
-    setTableData((_prevData) => {
-      const data = _prevData.data.filter(row => Number(row.id) !== Number(entityId));
-      return { ..._prevData, data };
+    
+    setRows((_prevData) => {
+      const data = _prevData.filter(row => Number(row.id) !== Number(entityId));
+      return data;
     });
   }
 
@@ -94,6 +119,7 @@ export default function Table() {
       <Search
         color={selected}
         onChange={(e) => table.setGlobalFilter(e.target.value)}
+        createRowHandler={createRowHandler}
       />
       <$TableBox width="100%">
         <$Table>
@@ -121,7 +147,7 @@ export default function Table() {
                   ))}
                 </tr>
                 {row.getIsExpanded() && (
-                  <tr>
+                  <tr>                    
                     <td colSpan={row.getVisibleCells().length}>
                       <RowDetail
                         entityId={row.original.id}
@@ -130,6 +156,7 @@ export default function Table() {
                         updateRowHandler={updateRowHandler}
                         deleteRowHandler={deleteRowHandler}
                         rowIndex={index}
+                        ref={template}
                       />
                     </td>
                   </tr>