+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.26 (2024-09-11)
+===============
+* [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/16)
+* Refactor table to have separate header and rows state
+* Implement api call to retrieve entity schema for headers
+* Create and implement logic for new entity creation
+
 0.25 (2024-09-09)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/15)